# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agent.pw API",
    "description": "API for managing credential profiles, credentials, tokens, and proxying authenticated requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.agent.pw"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Biscuit bearer token"
      }
    },
    "schemas": {
...
      "CredProfileDetail": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Unique profile identifier",
            "example": "linear"
          },
          "host": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Hostnames this profile applies to",
            "example": [
              "api.linear.app"
            ]
          },
          "path": {
            "type": "string",
            "description": "Path in the hierarchy",
            "example": "/"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable display name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Profile description"
          },
          "auth": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "Auth configuration (OAuth or headers)"
          }
        },
        "required": [
          "slug",
          "host",
          "path",
          "displayName",
          "description",
          "auth"
        ],
        "additionalProperties": false
      },
      "CreateCredProfileRequest": {
        "type": "object",
        "properties": {
          "host": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Hostnames this profile applies to"
          },
          "path": {
            "description": "Path for this profile (defaults to token path)",
            "example": "/",
            "type": "string"
          },
          "auth": {
            "description": "Auth configuration",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "managedOauth": {
            "description": "Managed OAuth config",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "displayName": {
            "description": "Human-readable display name",
            "type": "string"
          },
          "description": {
            "description": "Profile description",
            "type": "string"
          }
        },
        "required": [
          "host"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Profile not found"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
      "OkWithSlug": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "slug"
        ],
        "additionalProperties": false
      },
...
      "CreateCredentialRequest": {
        "type": "object",
        "properties": {
          "token": {
            "description": "API token or secret",
            "type": "string"
          },
          "headers": {
            "description": "Explicit header map to send on proxied requests",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "string"
            }
          },
          "host": {
            "description": "Target hostname for this credential",
            "example": "api.linear.app",
            "type": "string"
          },
          "profile": {
            "description": "Credential profile slug to derive the target host from",
            "example": "linear",
            "type": "string"
          },
          "path": {
            "description": "Full credential path (defaults to token path + name)",
            "example": "/orgs/ruzo/linear",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/credentials/{name}": {
      "put": {
        "operationId": "putCredentialsByName",
        "tags": [
          "credentials"
        ],
        "summary": "Store credential",
        "description": "Create or update a credential for a service. Provide either a token or explicit headers.",
        "responses": {
          "200": {
            "description": "Credential stored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    },
                    "name": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "name",
                    "path"
                  ],
                  "id": "CredentialStored"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "id": "CredentialError"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredentialRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ]
      },
      "delete": {
        "operationId": "deleteCredentialsByName",
        "tags": [
          "credentials"
        ],
        "summary": "Delete credential",
        "description": "Delete a credential by name plus host/path context.",
        "responses": {
          "200": {
            "description": "Credential deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Credential not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ]
      }
    },
...
    "/cred_profiles/{slug}": {
      "get": {
        "operationId": "getCredProfilesBySlug",
        "tags": [
          "cred_profiles"
        ],
        "summary": "Get credential profile",
        "description": "Get a credential profile by slug.",
        "responses": {
          "200": {
            "description": "Profile details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredProfileDetail"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ]
      },
      "put": {
        "operationId": "putCredProfilesBySlug",
        "tags": [
          "cred_profiles"
        ],
        "summary": "Create or update credential profile",
        "description": "Register a new credential profile or update an existing one by slug.",
        "responses": {
          "200": {
            "description": "Profile created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkWithSlug"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredProfileRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ]
      },
      "delete": {
        "operationId": "deleteCredProfilesBySlug",
        "tags": [
          "cred_profiles"
        ],
        "summary": "Delete credential profile",
        "description": "Remove a credential profile by slug.",
        "responses": {
          "200": {
            "description": "Profile deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ]
      }
    },
...
  }
}